+++ core/src/runtime/record.rs
@@ -5,7 +5,7 @@ use super::program::Program;
 use super::Opcode;
 use crate::alu::AluEvent;
 use crate::bytes::{ByteLookupEvent, ByteOpcode};
-use crate::cpu::CpuEvent;
+use crate::cpu::{CpuEvent, MemoryRecordEnum};
 use crate::field::event::FieldEvent;
 use crate::runtime::MemoryRecord;
 use crate::syscall::precompiles::edwards::EdDecompressEvent;
@@ -14,6 +14,8 @@ use crate::syscall::precompiles::keccak256::KeccakPermuteEvent;
 use crate::syscall::precompiles::sha256::{ShaCompressEvent, ShaExtendEvent};
 use crate::syscall::precompiles::{ECAddEvent, ECDoubleEvent};
 
+/// A record of the execution of a program. Contains event data for everything that happened during
+/// the execution of the segment.
 #[derive(Default, Clone, Debug)]
 pub struct ExecutionRecord {
     /// The index of the segment.
@@ -71,8 +73,8 @@ pub struct ExecutionRecord {
 
     pub k256_decompress_events: Vec<K256DecompressEvent>,
 
-    /// Information needed for global chips. This shouldn't really be in "Segment" but for
-    /// legacy reasons, we keep this information in this struct for now.
+    /// Information needed for global chips. This shouldn't really be here but for legacy reasons,
+    /// we keep this information in this struct for now.
     pub first_memory_record: Vec<(u32, MemoryRecord, u32)>,
     pub last_memory_record: Vec<(u32, MemoryRecord, u32)>,
     pub program_memory_record: Vec<(u32, MemoryRecord, u32)>,
@@ -220,4 +222,12 @@ impl ExecutionRecord {
             nb_k256_decompress_events: self.k256_decompress_events.len(),
         }
     }
+}
+
+#[derive(Debug, Copy, Clone, Default)]
+pub struct OpRecord {
+    pub a: Option<MemoryRecordEnum>,
+    pub b: Option<MemoryRecordEnum>,
+    pub c: Option<MemoryRecordEnum>,
+    pub memory: Option<MemoryRecordEnum>,
 }
